@@ -41,10 +41,11 @@
   "dependencies": {
     "crypto-random-string": "^1.0.0",
     "lodash": "^4.17.4",
-    "material-ui": "^0.18.6",
+    "material-ui": "^0.19.4",
     "moment": "^2.18.1",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
+    "react": "^16.1.0",
+    "react-dom": "^16.1.0",
+    "prop-types": "^15.6.0",
     "react-redux": "^5.0.5",
     "react-tap-event-plugin": "^2.0.1",
     "redux": "^3.7.2",
