@@ -1,8 +1,8 @@
-import React from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import TextField from 'material-ui/TextField';
 import { trim } from 'lodash';
 
-const { PropTypes, Component } = React;
 export default class TextDataField extends Component {
   static propTypes = {
     className: PropTypes.string,
